Adding control for random seed in theta perturbations
The random seed can now be set so that different iterations can be made for the same simulation. The use-case here is if you want to have statistically similar realizations of the same simulation, you can just change the random seed manually. If you just run the same simulation over and over, you will get the exact same result. This is useful for averaging over time in non-stationary flows (you can average over the realizations).